ajustes em todo backend

+++ boscov-backend/src/schemas/AvaliacaoSchema.ts
@@ -1,7 +1,7 @@
 import { z } from 'zod';
 export const avaliacaoSchema = z.object({
-    filmeId: z.number().int(),
-    nota: z.number().min(0).max(10),
-    comentario: z.string().optional(),
-  });
+  filmeId: z.number(),
+  nota: z.number().min(0).max(10),
+  comentario: z.string().optional()
+});
